Fix ContainerLeafNodeStructTreeValue
